Pre-calculate AttributeValue dependencies on changes
Previously, we were just-in-time calculating the graph for the
DependentValuesUpdate at the beginning of the job. This could take a
significant amount of time for non-trivial graphs.

Instead, whenver there is a change, involving the sources of data for an
AttributeValue in a Component, we calculate the input AttributeValues
used by every AttributeValue for that Component, so we can more quickly
return the graph starting from any atbitrary AttributeValue. When an
edge is made, connecting the ExternalProvider of one Component, to the
InternalProvider of another Component, the Component with the
InternalProvider is the one that has had sources change, and is the one
that is recalculated. In the case of an edge removal, it is still the
Component with the InternalProvider that has had sources change, and is
recalculated.

Because we are recalculating this next-hop graph information whenever
something in a Component has changed, writes are slower, but because
this information has already been generated in smaller chunks, getting
the dependency graph is significantly faster.