feat(sdf-server, dal): Exec variant def creates new variant, copies funcs
Instead of deleting the old schema variant and creating a new one, then attempting to migrate components, we just create a new schema variant on exec_variant_def and mark it as the default variant. Existing components are unchanged.

Component upgrades are coming next.

This is controlled by the "multiVariantEditing" flag, which is not yet in posthog (but can be easily turned on in the feature flags store).