Make rebases non-blocking and use work queue retention policy
This commit ensures the new engine uses the "work queue" retention
policy when using NATS Jetstream. As part of this change, we now make
rebases non-blocking by spawning them off into separate tasks. This is
unlikely to be the long term shape, but it's a step closer in the right
direction.

Signed-off-by: Nick Gerace <nick@systeminit.com>