feat(dal,sdf-server): migrate components to other schema variants
Adds the ability to take a component based on one schema variant and
migrate it to another schema variant without component deletion. Adds an
sdf route that facilitates this.

Frontend code not commited, resources have not been tested, and one
important feature is missing from this: the deletion and restoration of
edges to the new providers if they are compatible. But otherwise this
reliably "respins" a component onto a new variant.