chore(ci): set OpenTelemetry environment varible in integration tests
This change configures the OpenTelemetry collector endpoint for any
Cyclone servers running as part of an integration test. Prior to this
change (and given past work), Cyclone servers try in vain to connect to
a localhost collector only to fail. This causes a *lot* of stderr output
in the CI logs, even though this is not a failure at all.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>