fix(sdf): async routes should not commit on error by default
Committing in the error handler for async tasks means we are not rolling
our transactions back when a fatal error occurs. In the case of
workspace import, this is catastrophic, since we begin import by
clearing out the workspace entirely. A failed workspace import then
leaves the user with an empty workspace, destroying any data they
currently had.

Adds the ability to publish a WsEvent immediately, without requiring the
commit to drain the pending event queue.

Adds a standardized error handler function `handle_error` which
publishes the error event immediately and produces error telemetry.

The telemetry message for a failed async task looks like this now:

```
2024-02-22T17:18:10.901152Z ERROR ThreadId(47) sdf_server::server::service::async_route: async route '/api/pkg/install_pkg' error: could not find schema Docker Image with variant v1 for package component mysql
```

We should probably abstract more of the async behavior with a function
that takes an async `FnOnce() -> Result<T, E>` spawns the tokio task,
and handles the error itself. This will allow us to use the try (`?`)
operator in the async code.