Restore ability to update and list secrets in the new engine
This commit restores the ability to update and list secrets. Updating
secrets remains the same, but now the referential secrets on the
snapshot are updated. Listing secrets uses the same interfaces, but now
the graph is purely used and mimics the original logic (this first pass
is likely unoptimized as a result).

This is also the first time integration tests have been restored that
aren't part of the "new_engine" module! The secrets tests return. We
needed to expose "WorkspaceSignup" from the "dal-test" crate as a
result, since "WorkspaceSignup" is still in the "dal" on main. The
"test_harness" module has been partially restored as well.

Signed-off-by: Nick Gerace <nick@systeminit.com>