Fix content pair creation races by only writing out
Fix content pair creation races by only writing out and not returning
anything. We do not use long-running transactions with the content
store, and since there may be many users of the content store, we should
not be picky when writing out. Simply write out, do nothing if the pair
already exists, and move on.

Signed-off-by: Nick Gerace <nick@systeminit.com>
Co-authored-by: Paulo Cabral Sanz <paulo@systeminit.com>