Reimplement actions skeleton
Actions don't run yet, but they are enqueued/dequeued, batches are created and they get processed on apply.

It doesn't have tests yet, that's also a next step in a subsequent PR.

Also fixes a bunch of misc bugs that we found along the way, that's why it's being merged even if not feature complete.

Co-Authored-By: Nick Gerace <nick@systeminit.com>