chore(dal): use alternative to JoinSet Id to drop unstable requirement
This change tracks an alternate "task ID" (in this case a random `Ulid`)
which is used in the same `task_id_to_av_id` hash map logic, thus
removing the need for the unstable task `Id` type. In this way it's a
little easier to setup a development environment with editors that use
`cargo clippy` and `rust-analyzer` underneath.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>