feat(si-layer-cache): Make memory cache generic, store deserialized data
Moka has a strange requirement: it's Key and Value types have a
`'static` lifetime bound. This makes it tricky to capture it in a
generic type, since it means that type's generics also have to have a
static lifetime bound. We work around this by using a trait and
implementing that trait for all Cache<K, V>. A consequence is that the
Layer cache now has to take the concrete Moka cache type as a
dependency.

Co-Authored-By: Scott Prutton <scott@systeminit.com>