feat: add otelcol for integration tests
The 'RUSTC_BOOTSTRAP' var is used in conjunction with the buildkite-test-collector which must run using the
nightly compiler or using this bootstrapping flag to enable the json reporting option which has not yet made
it into the stable compiler as of 8th Mar 2024. This flag must not be used with the library compilation rule
above as it's generally intended to be used when compiling rust with older versions of rust.

The 'CI' var is used for the buildkite-test-collector which can't by default pick this var up in buildkite + buck2
setting it here has no impact on the local tests, it just allows the collector to run in active mode locally.
See https://buildkite.com/docs/test-analytics/rust-collectors for more info