chore(*): make all graph accesses async, remove need for lock management
This moves the lock on the snapshot graph inside the WorkspaceSnapshot,
so the caller never has to manage the lock. Instead, the lock is
controlled by the methods defined in the WorkspaceSnapshot
implementation. A consequence of this is that all methods on
WorkspaceSnapshot must be async, since the RwLock methods are async. But
we want this to happen anyway, since shrinking the snapshot means all
accesses will have to go through async boundaries to fetch node weights
from the caches.