fix(si-layer-cache): write generates the keys from the hash of the value
Generating the key before sending the data to the `write` method means
serializing twice, since we'll need to get the postcard bytes to
serialize the data (and then serialize the data in the write method).
Write should handle this. For non content-address keys we will need a
different interface.

Also adds some extensions to some types in `si-events` to ease
conversion from dal types.

Co-Authored-By: Scott Prutton <scott@systeminit.com>