fix(dal): restore old engine behavior when preparing function arguments
All prepared function arguments for a prototype function are gathered as
a mapping between function argument names and an *array* of values. In
the old engine, we always transformed that array of values to a single
value if the array was len() == 1. The new engine func binding
preparation special-cased input sockets function arguments to always
send an array, since this means that functions take take input socket
values as an input would not require special logic for handling the
scalar case and the array case (e.g., the si:normalizeToArray function).
However, our components are not built with this change in mind, and the
new behavior breaks some components slightly (for example EC2 Instance
sets the user data prop from the input socket with identity, so what we
write to the string prop turns into an array value).