Destroy ChangeSet struct
This commit removes the ChangeSet struct, moves the "force new"
functionality to ChangeSetPointer and removes relevant queries that use
ChangeSet.

This commit does not remove migrations, does not remove ChangeSetPk, nor
does it rename ChangeSetPointer to ChangeSet. This is the first step
towards fully removing the old ChangeSet concept and it should help
remove boilerplate for those working on the new engine.

Signed-off-by: Nick Gerace <nick@systeminit.com>