feat(dal,si-events): node weights off the graph
The node weights are now stored in layerdb. On the graph itself, a
"graph local node weight" stores only the address of the node weight and
the merkle tree hash. It was left on to allow calculating the hash using
the graph alone, but that is not actually possible, so could be removed
from the graph as well.

Co-Authored-By: Scott Prutton <scott@systeminit.com>