fix(dal): rebase must happen before dvu, but after commit
The rebase request must occur *after* a dal context has committed but
*before* any dependent value update job has run, since it's the
rebaser's responsibility to update the changeset pointer, and if the
dependent values job picks up the change set pointer before it has been
updated, the snapshot it uses might not contain the attribute values we
have asked it to update. This ensures we make the rebase request at the
correct time in commit (blocking_commit is already correct).