Update Component inputs/outputs to reflect when a Component is (un)marked as to_delete
Data should be able to flow between components following the rules
below:

| From Component `to_delete` | To Component `to_delete` | Data Propagates |
| -------------------------- | ------------------------ | --------------- |
| False                      | False                    | Yes             |
| False                      | True                     | Yes             |
| True                       | False                    | No              |
| True                       | True                     | Yes             |

Whenever a Component is marked (or un-marked) as `to_delete`, we need to
make sure that both its inputs, and all downstream Components' inputs
are updated to reflect the rules above. This allows us to both show the
effect of "deleting" a Component on the rest of the workspace, and keep
data that may be necessary to perform the delete action in Components
that are marked for deletion.