Update Rust dependencies since merging the new-engine into main
This commit updates every Rust-based dependency by its major, minor or
patch version by default. It also moves the `postcard` version
declaration from `dal` to the main `Cargo.toml` since the `alloc`
feature is a subset of the `use-std` feature.

Given the large scope of these changes, this commit does not seek to
change any Rust code or fixups barring a small change regarding
`EnumVariantNameds`. Exceptions to the goal of this commit include the
following:

- Hold `rustls`, `webpki-roots`, `tokio-vsock` and `yrs` to their minor
  versions due to API changes
- Hold `nix` and `ring` to their current versions due to needing
  buck2-specific changes
- Hold all hyper v1 dependencies to their minor versions due to a
  longer term effort needed to get us to hyper v1
- Hold `serde_yaml` to its current version due to being unmaintained
  (https://github.com/dtolnay/serde-yaml/releases/tag/0.9.34)

Signed-off-by: Nick Gerace <nick@systeminit.com>