Rename ChangeSetPointer to ChangeSet
This commit renames ChangeSetPointer to ChangeSet. It removes the old
"change_set" module in the dal and replaces it with the
"change_set_pointer" module. Events and status enums, structs and
functions are moved to two new submodules.

In addition, "force_changeset_pk" is replaced with "force_change_set_id"
in both sdf and the frontend.

Signed-off-by: Nick Gerace <nick@systeminit.com>