Annihilate all unsued migrations and queries in the dal
This is it. Since merging the new engine onto main, we have kept around
migrations and queries to help restore functionality in the system.
We're now at a point where we can destroy the bulk of our old migrations
and queries. You all served us well.

Additional or noteworthy changes:

- Move and rename all remaining queries to subdirectories based on which
  tables they reference
- Remove dead code related to migrations and queries
- Modify func-related and validation-related migrations to work since
  they are still used a hybrid fashion
- Make FuncBinding methods private that can be made private
- Fix "todo" for missing builtin workspace
- Fix "get_diff" integration test final assertion

Signed-off-by: Nick Gerace <nick@systeminit.com>