Replace FuncVariant with FuncKind and consolidate list func logic
This commit replaces FuncVariant with FuncKind. FuncKind is being used
in the new engine and FuncVariant is a very similar, but leftover, enum
from the func authoring experience in the old engine. These changes do
not guarantee that creating funcs will work, but listing and getting
funcs in the frontend should continue to work.

This commit also consolidates the "list_func" route logic into one
location in the dal. It is backed with an integration test as a result.

Misc changes:
- Add "kind" to the Func struct since it is already on the node weight
- Move FuncKind to its own submodule
- Remove unnecessary conversions of FuncBackendKind <> FuncVariant and
  hardcode FuncBackendKind in relevant routes
- Separate FuncKind and CustomizableFuncKind in the frontend (the latter
  is a subset of the former)
- Add ability to get default SchemaVariantId without fetching the
  entire Schema or SchemaVariant
- Replace usages of "transparent" in error types with populated fields
  as well as generally clean up error types
- Fix usage of "schema_name" in dal test helper

Signed-off-by: Nick Gerace <nick@systeminit.com>
Co-authored-by: Wendy Bujalski <wendy@systeminit.com>