Implement macro to enforce strong typing/direction in edges
Creates implement_add_edge_to macro that defines functions to specify strongly typed connections with the direction hardcoded.

This is a quick win implementation to solve a couple of footguns the system had, it does not intend on fixing all footguns. Since now is not the time to workshop a perfect solution this quick win felt like the way to go.

This unblocks copy pasting since it enumerates all connections in the system and makes it greppable.

Please try to use the implement_add_edge_to macro for new connections instead of workspace_snapshot.add_edge directly.

The current macro does not support all use cases, like category nodes, and definitely needs expansion.

It checks if edge weight kind in the function call matches the discriminant defined in the macro call.

Example:

Before:

    // Component (this) --> Schema Variant
    workspace_snapshot
        .add_edge(
            id, // Ulid
            EdgeWeight::new(change_set, EdgeWeightKind::new_use())?,
            schema_variant_id, // Ulid
        )
        .await?;

After:

    implement_add_edge_to!(
        source_id: ComponentId,
        destination_id: SchemaVariantId,
        add_fn: add_edge_to_schema_variant,
        discriminant: EdgeWeightKindDiscriminants::Use,
        result: ComponentResult,
    );

    // Enforces that the source is of type ComponentId and destination is of type SchemaVariantId
    // Runtime checks that EdgeWeightKind variant is Use
    Component::add_edge_to_schema_variant(
        ctx,
        component.id,
        schema_variant_id,
        EdgeWeightKind::new_use(),
    )
    .await?;

Feedback is needed here!

Benefits that this creates:

- Enforces connection direction
- Enforces connected objects types
- Runtime checks that the EdgeWeightKind passed is of the expected EdgeWeightKindDiscriminant
- Makes possible connections enumerable