feat(layerdb): replace sled with redb
Replacing sled, which was very much a "this will work for now" choice
with a different, simpler "this will work for now" choice, redb. It's
tiny, it's basically fancy lmdb, and it's being actively maintained. It also
seems fast enough for our purposes.

Also pinning refinery to a specific version until we upgrade rust.