bug(rebaser): only rebase one request at a time per change set
This PR fixes the issue where the rebaser would rebase things as fast as
it possibly could. Instead, it creates a queue for every change set ID
it sees, and only processes one message at a time for each change set.

It currently doesn't support removing those queues - once we see it, the
worker will stay resident forever. That's probably not a big deal, given
that we mostly have single users and the proliferation of change sets
isn't so high. But it's obviously not the optimal solution for long.

This is a stop-gap PR while Fletcher works on the long term solution,
which likely involves some clever use of Jetstream's tuning options to
ensure only one message can be in flight at a time, and that those
messages get replayed if needed.