Move working elements of func authoring to the dal
This PR moves working elements of Func authoring to the dal. We've
been reliant on integration tests to pave the way for stability in
the new engine. Moreover, the notion of "contexts" and how prototypes
behaves has radically changed in the new engine, so much of the Func
authoring experience needs to be refactored and restored piece by
piece. For these reasons, despite Func authoring potentially making
sense to be in sdf in the future, it will reside in the dal for now.

Primary changes:
- Move "get_func" functionality to the dal
- Move "create_func" functionality to the dal
- Move "save_func" functionality to the dal
- Restore ability to check if a Func is revertible
- Add optional "Func::get_by_id" and replace existing
  implementations with "Func::get_by_id_or_error"

Signed-off-by: Nick Gerace <nick@systeminit.com>