Find func associations for code generation and qualification funcs
This commit adds the ability to find func associations for code
generation and qualification funcs. It may have also done it for
attribute funcs, but that remains to be seen.

There's one integration test added as part of this commit, but it's
hardly enough for the authoring domain. Baby steps.

Signed-off-by: Nick Gerace <nick@systeminit.com>