Ignore deleted values in DVU
Ideally we should be dequeueing values that don't exist, but that's
a can of worms since we can't ensure it will happen, as removing some
edge may delete something because it gets garbage collected.

This should be ok, and not hide any bugs, and the warning is there so we
can be informed of further consequence. Right now it will pretty much
only happen in tests and copy pasting.

If the warning gets too verbose, we can remove it, or downgrade to a
debug statement, but let's see how it behaves first.

Co-Authored-By: Zack Hamm <zack@systeminit.com>