Restore Detatch Component
This PR restores the route to detach components from frames. The route takes a list of parent ids, so we can detach from many frames at once if needed. If no parent is passed in, we'll attempt to detach from the direct parent for that component. 

Follow up PR will include WS Events for these and adjustment to attachment to handle nested frames