Attach, Detach, and Nested Frames, Oh My!
This PR changes how we handle frames in the DAL. Previously, when connecting a Component to a Frame (up or down), we would create Attribute Prototype Arguments signifying which Input Socket was connected to which Output Socket for a given set of Components, in addition to creating a FrameContains edge between those components.  With this PR, we are relying solely on the FrameContains edge between Components and walking the lineage tree to determine the connection between Input and Output Sockets.  This means we no longer need an Attribute Prototype Argument between sockets, as the connection is 'inferred' from the lineage. As of this change, Input Sockets can only be driven through inference from a single Output Socket, though this can be expanded in the future as the need arises.

This PR also includes 6(!) new test exclusive schemas that can be used to write integration tests with interesting relationships. 'Even Lego' schemas have Input Sockets with even-numbered names, and Output Sockets with odd-numbered names. 'Odd Lego' schemas have Input Sockets with odd-numbered names and Output Sockets with even-numbered names.  There are small, medium, and large flavors of each Even/Odd Lego - allowing us to easily test the various permutations of inferred socket connections.  Small Legos have sockets with names "1", and "2", Medium has "1", "2", "3", "4", and Large Legos have "1","2","3","4","5",and "6".

I've also added an entry to the Component Debug panel to show when a Socket has an inferred connection.

Follow up PR will add WS Events for the frame attachment/detachment and fix up the UI to fill in the socket when it has an inferred connection.