feat(dal, sdf, web): Reintroduce PkgExport
Co-authored-by: Brit Myers <brit@systeminit.com>

This is now renamed as module export in the front end as we are dealing
with modules not packages.

This has been tested to export a package and then import it and it
works. The changes required to import both old and new format
packages is also working

We need to think about the tests needed to run this work as we want
to avoid needing to upload to the module index and then redownloading
the package